Refactor internal packages.
