+++ internal/pkg/common/range.go
@@ -1,4 +1,4 @@
-package mapping
+package common
 
 import "fmt"
